requirements.txt
streamlit
pandas
numpy
psycopg2-binary
plotly
